refactor: store 모듈 분리 (#33)
