+++ src/main/java/goorm/_44/config/WebConfig.java
@@ -1,4 +1,4 @@
-package goorm._44.config.security;
+package goorm._44.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
