+++ src/main/java/goorm/_44/controller/user/RefreshTokenController.java
@@ -1,7 +1,7 @@
 package goorm._44.controller.user;
 
-import goorm._44.config.api.ApiResult;
-import goorm._44.utils.JwtTokenProvider;
+import goorm._44.common.api.ApiResult;
+import goorm._44.common.jwt.JwtTokenProvider;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
